Model/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.47.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
cut-cross-entropy==25.1.1
datasets==3.6.0
diffusers==0.35.1
dill==0.3.8
docstring_parser==0.17.0
dotenv==0.9.9
eval_type_backport==0.2.2
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
hf-xet==1.1.10
hf_transfer==0.1.9
huggingface-hub==0.34.5
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.86.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
together==1.5.25
tokenizers==0.22.0
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
trl==0.23.0
typeguard==4.4.4
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
unsloth==2025.9.5
unsloth_zoo==2025.9.6
urllib3==2.5.0
uv==0.8.17
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
